Add content on setting labels for new fields in basic customisation section
Without defining labels for the new fields that were added, the user is left feeling they have done something wrong in adding the new step/fields because setting labels isn't covered in the inital setup.

To remedy this, added some extra content to the "Basic customisation" section, and moved the "content and i18n" section up to immediately follow.